Write parameterized union types, methods, and interfaces to separate file with go build tag
