+++ conjure/conjure.go
@@ -69,10 +69,14 @@ func GenerateOutputFiles(conjureDefinition spec.ConjureDefinition, cfg OutputCon
 		}
 		if len(pkg.Unions) > 0 {
 			unionFile := newJenFile(pkg, def)
+			go1Dot18UnionFile := newJenFile(pkg, def)
+			go1Dot18UnionFile.Comment("//go:build go1.18")
 			for _, union := range pkg.Unions {
 				writeUnionType(unionFile.Group, union, cfg.GenerateFuncsVisitor)
+				writeUnionTypeWithGenerics(go1Dot18UnionFile.Group, union)
 			}
 			files = append(files, newGoFile(filepath.Join(pkg.OutputDir, "unions.conjure.go"), unionFile))
+			files = append(files, newGoFile(filepath.Join(pkg.OutputDir, "unions.conjure_go1.18.go"), go1Dot18UnionFile))
 		}
 		if len(pkg.Errors) > 0 {
 			errorFile := newJenFile(pkg, def)
