+++ integration_test/testgenerated/objects/api/unions.conjure.go
@@ -177,36 +177,6 @@ type ExampleUnionVisitorWithContext interface {
 	VisitUnknownWithContext(ctx context.Context, typeName string) error
 }
 
-type ExampleUnionWithT[T any] ExampleUnion
-
-func (u *ExampleUnionWithT[T]) Accept(ctx context.Context, v ExampleUnionVisitorWithT[T]) (T, error) {
-	switch u.typ {
-	default:
-		if u.typ == "" {
-			var result T
-			return result, fmt.Errorf("invalid value in union type")
-		}
-		return v.VisitUnknown(ctx, u.typ)
-	case "str":
-		return v.VisitStr(ctx, *u.str)
-	case "strOptional":
-		var strOptional *string
-		if u.strOptional != nil {
-			strOptional = *u.strOptional
-		}
-		return v.VisitStrOptional(ctx, strOptional)
-	case "other":
-		return v.VisitOther(ctx, *u.other)
-	}
-}
-
-type ExampleUnionVisitorWithT[T any] interface {
-	VisitStr(ctx context.Context, v string) (T, error)
-	VisitStrOptional(ctx context.Context, v *string) (T, error)
-	VisitOther(ctx context.Context, v int) (T, error)
-	VisitUnknown(ctx context.Context, typ string) (T, error)
-}
-
 func NewExampleUnionFromStr(v string) ExampleUnion {
 	return ExampleUnion{typ: "str", str: &v}
 }
